+++ internal/dataplane/kong_client_test.go
@@ -1,0 +1,107 @@
+package dataplane
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+	corev1 "k8s.io/api/core/v1"
+	netv1 "k8s.io/api/networking/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime/schema"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	"github.com/kong/kubernetes-ingress-controller/v2/internal/dataplane/failures"
+)
+
+func TestUniqueObjects(t *testing.T) {
+	t.Log("generating some objects to test the de-duplication of objects")
+	ing1 := &netv1.Ingress{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace:  corev1.NamespaceDefault,
+			Name:       "test-ingress-1",
+			Generation: 1,
+		},
+	}
+	ing1.SetGroupVersionKind(ingGVK)
+	ing2 := &netv1.Ingress{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace:  corev1.NamespaceDefault,
+			Name:       "test-ingress-2",
+			Generation: 1,
+		},
+	}
+	ing2.SetGroupVersionKind(ingGVK)
+	ing3 := &netv1.Ingress{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace:  "other-namespace",
+			Name:       "test-ingress-1",
+			Generation: 1,
+		},
+	}
+	ing3.SetGroupVersionKind(ingGVK)
+	ing4 := &netv1.Ingress{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace:  "other-namespace",
+			Name:       "test-ingress-2",
+			Generation: 1,
+		},
+	}
+	ing4.SetGroupVersionKind(ingGVK)
+
+	testCases := []struct {
+		name         string
+		reportedObjs []client.Object
+		failedObjs   [][]client.Object
+		uniqueObjs   []client.Object
+	}{
+		{
+			name:         "no failures",
+			reportedObjs: []client.Object{ing1, ing2},
+			uniqueObjs:   []client.Object{ing1, ing2},
+		},
+		{
+			name:         "has failures",
+			reportedObjs: []client.Object{ing1, ing3},
+			failedObjs: [][]client.Object{
+				{ing1},
+				{ing4},
+			},
+			uniqueObjs: []client.Object{ing1, ing3, ing4},
+		},
+		{
+			name:         "one object in multiple failures",
+			reportedObjs: []client.Object{ing1, ing2},
+			failedObjs: [][]client.Object{
+				{ing3},
+				{ing2, ing3},
+			},
+			uniqueObjs: []client.Object{ing1, ing2, ing3},
+		},
+	}
+
+	for _, tc := range testCases {
+		tc := tc
+		t.Run(tc.name, func(t *testing.T) {
+			translationFailures := []failures.ResourceFailure{}
+			for _, failedObjs := range tc.failedObjs {
+				translationFailure, err := failures.NewResourceFailure(
+					"for test", failedObjs...,
+				)
+				require.NoError(t, err)
+				translationFailures = append(translationFailures, translationFailure)
+			}
+			uniqueObjs := uniqueObjects(tc.reportedObjs, translationFailures)
+			require.Len(t, uniqueObjs, len(tc.uniqueObjs))
+			require.ElementsMatch(t, tc.uniqueObjs, uniqueObjs)
+		})
+	}
+}
+
+// initialized objects don't have GVK's, so we fake those for unit tests.
+var (
+	ingGVK = schema.GroupVersionKind{
+		Group:   "networking.k8s.io",
+		Version: "v1",
+		Kind:    "Ingress",
+	}
+)